requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
python-dotenv==1.0.0
better-exceptions==0.3.2
numpy==1.24.3
torch==2.1.2
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
pyjwt==2.8.0
cryptography
httpx==0.25.2